Update wagmi generation
